feat: handlers and renders are refactored
