+++ src/handlers/storageList/storageListCreate.cpp
@@ -1,0 +1,24 @@
+#include "storageListCreate.hpp"
+#include "../../render/storageList/storageListRender.hpp"
+
+namespace handlers::storageListCreate {
+
+using namespace render::storageList;  
+using namespace cookcookhnya::forHandlers;
+
+void createStorage(StorageCreationEnterName&, MessageRef m, BotRef bot, SMRef stateManager, BackendApiRef api) { // BackendProvider bkn
+    api.create(m.from->id, models::storage::StorageCreateBody{m.text}); // Create storage bpdy with new name
+    stateManager.put(StorageList{});
+    renderStorageList(m.from->id, m.chat->id, bot, api);
+};
+
+void cancelStorageCreation(StorageCreationEnterName&, CallbackQueryRef cq, BotRef bot, SMRef stateManager, BackendApiRef api) {
+    bot.answerCallbackQuery(cq.id);
+    if (cq.data == "StorageCreateCancel") { // Here compare with data in button which was pressed (data was put in
+                                            // renderStorageCreate)
+        stateManager.put(StorageList{});
+        renderStorageList(cq.from->id, cq.message->chat->id, bot, api);
+    }
+};
+
+}  // namespace handlers::storageListCreate
