+++ src/handlers/storageView/storageViewMembers.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../typeRefs.hpp"
+
+namespace handlers::storageViewMembers {
+using namespace cookcookhnya::forHandlers;
+
+extern void storageMemberViewButtonCallback(StorageMemberView& state, CallbackQueryRef cq, BotRef bot, SMRef stateManager,  BackendApiRef api);
+}  // namespace handlers::storageViewMembers
