+++ src/render/ingredientsView/ingredientsRender.cpp
@@ -1,0 +1,17 @@
+#include "ingredientsRender.hpp"
+
+namespace render::viewIngredients{
+using namespace cookcookhnya::render;
+
+void renderIngredientsList(StorageId storageId, UserId userId, ChatId chatId, BotRef bot) {
+    // auto ad = backendEx.getUserStorages(userId)[0];
+    // std::string resultStr ="";
+    // for (int i=0;i<ad.getContent(123).size();i++){
+    //     resultStr += std::format("{} \n", ad.getContent(123)[i]);
+    // }
+    //     bot.sendMessage(chatId,
+    //                 resultStr);
+
+    return;
+}
+}  // namespace render::viewIngredients
