+++ src/render/storageList/createStorageRender.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../typeForRender.hpp"
+
+namespace render::createStorage{
+using namespace cookcookhnya::render;
+
+extern void renderStorageCreate(ChatId chatId, BotRef bot, UserId userId);
+} // namespace render::createStorage
